requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
dnspython==2.5.0
Faker==22.6.0
frozenlist==1.4.1
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.1
six==1.16.0
typing_extensions==4.9.0
yarl==1.9.4